create location modal ui
