+++ mfg-kanban-frontend/src/lib/components/modals/ManageLocationModal.svelte
@@ -1,9 +1,8 @@
 <script lang="ts">
 	import type { LocationType } from '$lib/types';
-	import { onMount } from 'svelte';
 
-	import Sortable from 'sortablejs';
 	import ModalTemplate from './ModalTemplate.svelte';
+	import LocationReorder from './LocationReorder.svelte';
 
 	export let showModal: boolean;
 	// temporary, should get locations, parts from fetch
@@ -13,28 +12,13 @@
 	let locationSequence: number;
 
 	async function submit() {}
-
-	let sortableEle: HTMLElement;
-	let sortableObj: Sortable;
-
-	onMount(() => {
-		sortableObj = Sortable.create(sortableEle, {
-			group: 'location',
-			animation: 150,
-			ghostClass: 'blue-background-class',
-			onEnd: function (evt) {
-				// Need to figure out how update database
-				console.log('sort ended');
-			}
-		});
-	});
 </script>
 
 <ModalTemplate bind:showModal modalTitle="Mange Location">
 	<div class="my-2 mx-2">
-		<div class="flex flex-row">
+		<div class="flex flex-col">
 			<form on:submit|preventDefault={submit}>
-				<div class="grid grid-cols-1 gap-6">
+				<div class="flex flex-row my-2 gap-2">
 					<div class="block">
 						<span class="text-gray-600">Name</span>
 						<input
@@ -56,21 +40,13 @@
 							min="1"
 						/>
 					</div>
+					<button
+						class="py-2 mt-6 rounded-md px-4 bg-blue-600 text-white hover:bg-white hover:text-blue-600 hover:border-blue-600 hover:border"
+						type="submit">Create</button
+					>
 				</div>
-				<button
-					class="my-2 py-2 px-4 rounded-full bg-blue-600 text-white hover:bg-white hover:text-blue-600 hover:border-blue-600 hover:border"
-					type="submit">Create</button
-				>
 			</form>
-			<div class="flex flex-col">
-				<div bind:this={sortableEle} class="">
-					{#each Locations as Location}
-						<div class="font-semibold border shadow-md rounded-lg">
-							{Location.name}
-						</div>
-					{/each}
-				</div>
-			</div>
+			<LocationReorder {Locations} />
 		</div>
 	</div>
 </ModalTemplate>
